Remove duplicated styles for modals
