+++ src/modals/roadmap/Header.tsx
@@ -14,16 +14,17 @@ import {
 import { Ticket } from "@typings/roadmap";
 
 import {
-  ModalHeaderWrapper,
-  ModalHeaderIcon,
+  ModalHeaderLinks,
+  ModalHeaderLink,
+  ModalHeaderLinkButton,
+} from "@components/shared/Modal/styled";
+import {
   ModalHeaderMeta,
+  ModalHeaderIcon,
   ModalHeaderMetaCapability,
   ModalHeaderMetaTicketNumber,
   ModalHeaderDueDate,
   ModalHeaderTitle,
-  ModalHeaderLinks,
-  ModalHeaderLink,
-  ModalHeaderLinkButton,
 } from "./styled";
 
 export interface ModalHeaderProps {
@@ -41,40 +42,38 @@ export const ModalHeader: React.FunctionComponent<ModalHeaderProps> = ({
     githubUrl,
   },
 }) => (
-  <ModalHeaderWrapper>
-    <div>
-      <ModalHeaderIcon>
-        <CapabilitySvg capability={capability.id} />
-      </ModalHeaderIcon>
-      <ModalHeaderMeta>
-        <ModalHeaderMetaCapability>
-          {capability.displayName}
-        </ModalHeaderMetaCapability>
-        <ModalHeaderMetaTicketNumber>
-          <TicketNumber number={nr} />
-        </ModalHeaderMetaTicketNumber>
-      </ModalHeaderMeta>
-      <ModalHeaderDueDate>
-        <Icon iconName="calendar-alt" iconPrefix="far" />
-        <DueDate date={dueDate} futurePlanned={release.title === "Future"} />
-      </ModalHeaderDueDate>
-      <ModalHeaderTitle>{title}</ModalHeaderTitle>
-      <ModalHeaderLinks>
-        <ModalHeaderLink>
-          <Link.External to={zenHubUrl}>
-            <ModalHeaderLinkButton>
-              <FormattedMessage id="roadmap.modal.viewInZenHub" />
-            </ModalHeaderLinkButton>
-          </Link.External>
-        </ModalHeaderLink>
-        <ModalHeaderLink>
-          <Link.External to={githubUrl}>
-            <ModalHeaderLinkButton>
-              <FormattedMessage id="roadmap.modal.viewInGitHub" />
-            </ModalHeaderLinkButton>
-          </Link.External>
-        </ModalHeaderLink>
-      </ModalHeaderLinks>
-    </div>
-  </ModalHeaderWrapper>
+  <>
+    <ModalHeaderIcon>
+      <CapabilitySvg capability={capability.id} />
+    </ModalHeaderIcon>
+    <ModalHeaderMeta>
+      <ModalHeaderMetaCapability>
+        {capability.displayName}
+      </ModalHeaderMetaCapability>
+      <ModalHeaderMetaTicketNumber>
+        <TicketNumber number={nr} />
+      </ModalHeaderMetaTicketNumber>
+    </ModalHeaderMeta>
+    <ModalHeaderDueDate>
+      <Icon iconName="calendar-alt" iconPrefix="far" />
+      <DueDate date={dueDate} futurePlanned={release.title === "Future"} />
+    </ModalHeaderDueDate>
+    <ModalHeaderTitle>{title}</ModalHeaderTitle>
+    <ModalHeaderLinks>
+      <ModalHeaderLink>
+        <Link.External to={zenHubUrl}>
+          <ModalHeaderLinkButton>
+            <FormattedMessage id="roadmap.modal.viewInZenHub" />
+          </ModalHeaderLinkButton>
+        </Link.External>
+      </ModalHeaderLink>
+      <ModalHeaderLink>
+        <Link.External to={githubUrl}>
+          <ModalHeaderLinkButton>
+            <FormattedMessage id="roadmap.modal.viewInGitHub" />
+          </ModalHeaderLinkButton>
+        </Link.External>
+      </ModalHeaderLink>
+    </ModalHeaderLinks>
+  </>
 );
